+++ app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,0 +1,82 @@
+package com.techcourse.service;
+
+import com.techcourse.config.DataSourceConfig;
+import com.techcourse.domain.User;
+import java.sql.Connection;
+import java.sql.SQLException;
+import javax.sql.DataSource;
+import org.springframework.dao.DataAccessException;
+import org.springframework.jdbc.datasource.DataSourceUtils;
+
+public class TxUserService implements UserService {
+
+    private final UserService userService;
+
+    public TxUserService(final UserService userService) {
+        this.userService = userService;
+    }
+
+    @Override
+    public User findById(final long id) {
+        final DataSource dataSource = DataSourceConfig.getInstance();
+        final Connection connection = DataSourceUtils.getConnection(dataSource);
+        try {
+            connection.setAutoCommit(false);
+
+            final User user = userService.findById(id);
+
+            connection.commit();
+
+            return user;
+        } catch (SQLException | DataAccessException e) {
+            rollback(connection);
+            throw new DataAccessException(e);
+        } finally {
+            DataSourceUtils.releaseConnection(connection, dataSource);
+        }
+    }
+
+    @Override
+    public void insert(final User user) {
+        final DataSource dataSource = DataSourceConfig.getInstance();
+        final Connection connection = DataSourceUtils.getConnection(dataSource);
+        try {
+            connection.setAutoCommit(false);
+
+            userService.insert(user);
+
+            connection.commit();
+        } catch (SQLException | DataAccessException e) {
+            rollback(connection);
+            throw new DataAccessException(e);
+        } finally {
+            DataSourceUtils.releaseConnection(connection, dataSource);
+        }
+    }
+
+    @Override
+    public void changePassword(final long id, final String newPassword, final String createBy) {
+        final DataSource dataSource = DataSourceConfig.getInstance();
+        final Connection connection = DataSourceUtils.getConnection(dataSource);
+        try {
+            connection.setAutoCommit(false);
+
+            userService.changePassword(id, newPassword, createBy);
+
+            connection.commit();
+        } catch (SQLException | DataAccessException e) {
+            rollback(connection);
+            throw new DataAccessException(e);
+        } finally {
+            DataSourceUtils.releaseConnection(connection, dataSource);
+        }
+    }
+
+    private void rollback(final Connection connection) {
+        try {
+            connection.rollback();
+        } catch (SQLException e) {
+            throw new DataAccessException();
+        }
+    }
+}
